statistics_storage/requirements.txt
flask
grpcio
protobuf
redis
